♻️ refactor(cli): extract runWithConcurrency into shared module
