+++ apps/cli/src/import-wiki/index.ts
@@ -6,6 +6,7 @@ import {
   WIKI_TAGS,
   type WikiTag,
 } from "@howardism/article-contract";
+import { runWithConcurrency } from "../concurrency.ts";
 import { generateHeroImage } from "./codex.ts";
 import { type ArticleMeta, emitArticle } from "./emit.ts";
 import { buildManifests, writeManifests } from "./pages/manifests.ts";
@@ -430,30 +431,6 @@ async function tryParseWikiLog(
     }
     throw err;
   }
-}
-
-async function runWithConcurrency<T, R>(
-  items: T[],
-  concurrency: number,
-  worker: (item: T) => Promise<R>
-): Promise<R[]> {
-  const results: R[] = new Array(items.length);
-  let nextIndex = 0;
-  const workers = Array.from(
-    { length: Math.min(concurrency, items.length) },
-    async () => {
-      while (true) {
-        const i = nextIndex;
-        nextIndex += 1;
-        if (i >= items.length) {
-          return;
-        }
-        results[i] = await worker(items[i]);
-      }
-    }
-  );
-  await Promise.all(workers);
-  return results;
 }
 
 function parseOptions(): RunOptions {
